-исправление замечаний по auth app
