+++ app/src/main/java/com/example/sandboxbank/auth/data/dto/LoginRequest.kt
@@ -1,4 +1,4 @@
-package com.example.sandboxbank.data.dto.login
+package com.example.sandboxbank.auth.data.dto
 
 data class LoginRequest(
     val email: String,
