+++ app/src/main/java/com/example/sandboxbank/auth/data/storage/SecureStorage.kt
@@ -1,4 +1,4 @@
-package com.example.sandboxbank.data.storage
+package com.example.sandboxbank.auth.data.storage
 
 class SecureStorage {
     fun saveToken(accessToken: String) {
