+++ app/src/main/java/com/example/sandboxbank/auth/ui/screen/AuthScreen.kt
@@ -35,9 +35,9 @@ import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.example.sandboxbank.domain.model.ResultAuthState
+import com.example.sandboxbank.auth.domain.model.ResultAuthState
 import com.example.sandboxbank.mode.LightColorPalette
-import com.example.sandboxbank.ui.auth.viewmodel.AuthViewModel
+import com.example.sandboxbank.auth.ui.viewmodel.AuthViewModel
 
 
 @Composable
