--- app/src/main/java/com/example/sandboxbank/data/RetrofitClient.kt
@@ -1,17 +1,0 @@
-package com.example.sandboxbank.data
-
-import com.example.sandboxbank.data.dto.login.LoginRequest
-import com.example.sandboxbank.data.dto.login.LoginResponse
-import com.example.sandboxbank.data.dto.register.RegisterRequest
-import com.example.sandboxbank.data.dto.register.RegisterResponse
-import com.example.sandboxbank.data.repository.api.NetworkClient
-
-class RetrofitClient: NetworkClient {
-    override suspend fun register(registerRequest: RegisterRequest): RegisterResponse {
-        TODO("Not yet implemented")
-    }
-
-    override suspend fun login(loginRequest: LoginRequest): LoginResponse {
-        TODO("Not yet implemented")
-    }
-}
